move classes to dataset.py
